+++ src/main/java/mx/edu/utez/model/role/BeanRole.java
@@ -1,22 +1,22 @@
 package mx.edu.utez.model.role;
 
 public class BeanRole {
-    private short id;
+    private int id;
     private String description;
 
     public BeanRole() {
     }
 
-    public BeanRole(short id, String description) {
+    public BeanRole(int id, String description) {
         this.id = id;
         this.description = description;
     }
 
-    public short getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(short id) {
+    public void setId(int id) {
         this.id = id;
     }
